Adicionando exceções

+++ src/main/java/dev/gustavosdaniel/projetospringboot/service/ProdutoService.java
@@ -1,5 +1,6 @@
 package dev.gustavosdaniel.projetospringboot.service;
 
+import dev.gustavosdaniel.projetospringboot.exception.RecursoNaoEncontradoException;
 import dev.gustavosdaniel.projetospringboot.model.Produto;
 import dev.gustavosdaniel.projetospringboot.repository.ProdutoRepository;
 import org.springframework.stereotype.Service;
@@ -20,8 +21,9 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> buscarPorId(Long id) { // OPTIONAL É USADO POR CONTA QUE PODE EXISTIR UM UM ID OU NÃO POR ISSO USAR OPTIONAL
-        return produtoRepository.findById(id); // RETORNE O ID QUE FOI PESQUISADO
+    public Produto buscarPorId(Long id) { // OPTIONAL É USADO POR CONTA QUE PODE EXISTIR UM UM ID OU NÃO POR ISSO USAR OPTIONAL
+        return produtoRepository.findById(id).orElseThrow(() -> new RecursoNaoEncontradoException("Produto com ID " +id+ " não encontrado")); // RETORNE O ID QUE FOI PESQUISADO
+        // Caso não achar esse ID do produto vai retornar essa mensagem
     }
 
     public Produto salvarProduto(Produto produto) {
@@ -29,6 +31,12 @@ public class ProdutoService {
     }
 
     public void deletarProdutoPorId(Long id) { //TEM QUE SER VOID (VAZIO) POIS NÃO VAI RETORNAR NADA SO VAI APAGAR
+
+        if (!produtoRepository.existsById(id)) { // ESSE ! SIGINIFICA NEGAÇÃO
+            throw new RecursoNaoEncontradoException("Produto com ID "+id +" não encontrado.");
+        }
         produtoRepository.deleteById(id);
     }
+
+
 }
